requirements.txt
openai>=1.40.0
sqlglot>=20.0.0
python-dotenv>=1.0.1
